+++ RyuSocks/Packets/Packet.cs
@@ -34,6 +34,24 @@ namespace RyuSocks.Packets
         /// </summary>
         public abstract void Validate();
 
+        /// <summary>
+        /// Check whether the structure of the packet is valid.
+        /// This method calls <see cref="Validate"/> internally, but doesn't throw the exception on failure
+        /// and returns a <see langword="bool"/> instead.
+        /// </summary>
+        public bool IsValid()
+        {
+            try
+            {
+                Validate();
+            }
+            catch
+            {
+                return false;
+            }
+
+            return true;
+        }
 
         protected Packet() { }